+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Result status comparison
+- Raise error on extra args/kwargs
 
 ## [0.221.0-rc.0] - 2023-04-17